+++ src/plugin/core/net/i_ws_module.h
@@ -13,7 +13,7 @@
 class IWSModule : public IModule {
   public:
     template <typename BaseType>
-    bool AddReceiveCallBack(const int msg_id, BaseType *pBase, void (BaseType::*handleReceiver)(const socket_t, const int, const char *, const uint32_t)) {
+    bool AddReceiveCallBack(const uint32_t msg_id, BaseType *pBase, void (BaseType::*handleReceiver)(const socket_t, const uint32_t, const char *, const uint32_t)) {
         NET_RECEIVE_FUNCTOR functor =
             std::bind(handleReceiver, pBase, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4);
         NET_RECEIVE_FUNCTOR_PTR functorPtr(new NET_RECEIVE_FUNCTOR(functor));
@@ -22,7 +22,7 @@ class IWSModule : public IModule {
     }
 
     template <typename BaseType>
-    bool AddReceiveCallBack(BaseType *pBase, void (BaseType::*handleReceiver)(const socket_t, const int, const char *, const uint32_t)) {
+    bool AddReceiveCallBack(BaseType *pBase, void (BaseType::*handleReceiver)(const socket_t, const uint32_t, const char *, const uint32_t)) {
         NET_RECEIVE_FUNCTOR functor =
             std::bind(handleReceiver, pBase, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4);
         NET_RECEIVE_FUNCTOR_PTR functorPtr(new NET_RECEIVE_FUNCTOR(functor));
@@ -43,8 +43,8 @@ class IWSModule : public IModule {
     // as server
     virtual int Listen(const unsigned int nMaxClient, const unsigned short nPort, const int nCpuCount, const uint32_t expand_buffer_size) = 0;
 
-    virtual void RemoveReceiveCallBack(const int msg_id) = 0;
-    virtual bool AddReceiveCallBack(const int msg_id, const NET_RECEIVE_FUNCTOR_PTR &cb) = 0;
+    virtual void RemoveReceiveCallBack(const uint32_t msg_id) = 0;
+    virtual bool AddReceiveCallBack(const uint32_t msg_id, const NET_RECEIVE_FUNCTOR_PTR &cb) = 0;
     virtual bool AddReceiveCallBack(const NET_RECEIVE_FUNCTOR_PTR &cb) = 0;
     virtual bool AddEventCallBack(const NET_EVENT_FUNCTOR_PTR &cb) = 0;
 
@@ -53,8 +53,8 @@ class IWSModule : public IModule {
     virtual bool SendData(const std::string &msg, const socket_t sock, const bool text = true) = 0;
     virtual bool SendDataToAllClient(const std::string &msg, const bool text = true) = 0;
 
-    virtual bool SendMsg(const int16_t msg_id, const char *msg, const size_t len, const socket_t sock) = 0;
-    virtual bool SendPBMsg(const uint16_t msg_id, const google::protobuf::Message &xData, const socket_t sock) = 0;
+    virtual bool SendMsg(const int32_t msg_id, const char *msg, const size_t len, const socket_t sock) = 0;
+    virtual bool SendPBMsg(const uint32_t msg_id, const google::protobuf::Message &xData, const socket_t sock) = 0;
 
     virtual INet *GetNet() = 0;
 };
